refactor: 데이터 처리 함수 service.js로 분리
